requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
defusedxml==0.7.1
idna==3.7
jira==3.8.0
numpy==2.0.1
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
pillow==10.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
